Refactored: Behavior Tree File Renaming and Restructuring
