+++ modules/base_bt_nodes.py
@@ -1,8 +1,6 @@
 from enum import Enum
-import math
-import random
 # BT Node List
-class BehaviorTreeList:
+class BTNodeList:
     CONTROL_NODES = [        
         'Sequence',
         'Fallback'
@@ -10,9 +8,7 @@ class BehaviorTreeList:
 
     ACTION_NODES = [
         'LocalSensingNode',
-        'DecisionMakingNode',
-        'TaskExecutingNode',
-        'ExplorationNode'
+        'DecisionMakingNode'
     ]
 
     CONDITION_NODES = [
@@ -76,13 +72,6 @@ class SyncAction(Node):
 # Load additional configuration and import decision-making class dynamically
 import importlib
 from modules.utils import config
-from plugins.my_decision_making_plugin import *
-target_arrive_threshold = config['tasks']['threshold_done_by_arrival']
-task_locations = config['tasks']['locations']
-sampling_freq = config['simulation']['sampling_freq']
-sampling_time = 1.0 / sampling_freq  # in seconds
-agent_max_random_movement_duration = config.get('agents', {}).get('random_exploration_duration', None)
-
 decision_making_module_path = config['decision_making']['plugin']
 module_path, class_name = decision_making_module_path.rsplit('.', 1)
 decision_making_module = importlib.import_module(module_path)
@@ -114,53 +103,4 @@ class DecisionMakingNode(SyncAction):
         else:                        
             return Status.SUCCESS
 
-
-# Task executing node
-class TaskExecutingNode(SyncAction):
-    def __init__(self, name, agent):
-        super().__init__(name, self._execute_task)
-
-    def _execute_task(self, agent, blackboard):        
-        assigned_task_id = blackboard.get('assigned_task_id')        
-        if assigned_task_id is not None:
-            agent_position = agent.position
-            next_waypoint = agent.tasks_info[assigned_task_id].position
-            # Calculate norm2 distance
-            distance = math.sqrt((next_waypoint[0] - agent_position[0])**2 + (next_waypoint[1] - agent_position[1])**2)
-            
-            assigned_task_id = blackboard.get('assigned_task_id')
-            if distance < agent.tasks_info[assigned_task_id].radius + target_arrive_threshold: # Agent reached the task position                                
-                if agent.tasks_info[assigned_task_id].completed:  # 이렇게 먼저 해줘야 중복해서 task_amount_done이 올라가지 않는다.                  
-                    return Status.SUCCESS
-                agent.tasks_info[assigned_task_id].reduce_amount(agent.work_rate)
-                agent.update_task_amount_done(agent.work_rate)  # Update the amount of task done                
-
-            # Move towards the task position
-            agent.follow(next_waypoint)
-
-        return Status.RUNNING
-
-
-# Exploration node
-class ExplorationNode(SyncAction):
-    def __init__(self, name, agent):
-        super().__init__(name, self._random_explore)
-        self.random_move_time = float('inf')
-        self.random_waypoint = (0, 0)
-
-    def _random_explore(self, agent, blackboard):
-        # Move towards a random position
-        if self.random_move_time > agent_max_random_movement_duration:
-            self.random_waypoint = self.get_random_position(task_locations['x_min'], task_locations['x_max'], task_locations['y_min'], task_locations['y_max'])
-            self.random_move_time = 0 # Initialisation
-        
-        blackboard['random_waypoint'] = self.random_waypoint        
-        self.random_move_time += sampling_time   
-        agent.follow(self.random_waypoint)         
-        return Status.RUNNING
-        
-    def get_random_position(self, x_min, x_max, y_min, y_max):
-        pos = (random.randint(x_min, x_max),
-                random.randint(y_min, y_max))
-        return pos
